Fix ( Controller - Service - Repository )
1. To Save List is finished ( repo.save -> repo.saveAll )
2. Add Menu validation
3. Delete menu history page

+++ src/main/java/com/lunch/repository/HistoryRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.lunch.domain.History;
 
 @Repository
-public interface HistoryRepository extends JpaRepository<History, Long>{
+public interface HistoryRepository extends JpaRepository<History, Integer>{
 	
 }
